Develop upstream sync 220912

+++ tensorflow/compiler/xla/service/gpu/nccl_all_to_all_thunk.cc
@@ -106,7 +106,8 @@ Status RunAllToAll(bool has_split_dimension,
           auto dtype_and_multiplier,
           ToNcclDataTypeAndCountMultiplier(buffer.element_type));
       ncclDataType_t dtype = dtype_and_multiplier.first;
-      int element_count = buffer.element_count * dtype_and_multiplier.second;
+      int64_t element_count =
+          buffer.element_count * dtype_and_multiplier.second;
 
       TF_RET_CHECK(element_count % num_participants == 0)
           << "Buffer was not an exact multiple of the number of participants.";
