+++ tensorflow/compiler/xla/service/gpu/nccl_collective_permute_thunk.cc
@@ -186,7 +186,7 @@ Status RunCollectivePermute(
   TF_ASSIGN_OR_RETURN(auto dtype_and_multiplier,
                       ToNcclDataTypeAndCountMultiplier(buffer.element_type));
   ncclDataType_t dtype = dtype_and_multiplier.first;
-  int element_count = buffer.element_count * dtype_and_multiplier.second;
+  int64_t element_count = buffer.element_count * dtype_and_multiplier.second;
 
   se::gpu::GpuStreamHandle gpu_stream = se::gpu::AsGpuStreamValue(&stream);
